zcash_pool_migration: extract the persisted vocabulary into `model`
`engine.rs` held four separable things at once: the persisted vocabulary,
the trait seams, the plan types, and the operations. Because the vocabulary
lived there, every module that needed a `MigrationTransferId` had to import
from the engine, while the engine imported every module for its
combinatorics. That produced two module cycles, `engine <-> state` and
`engine <-> signing_rounds`, neither of which reflected a real dependency.

Move the vocabulary out to `model`: the transfer id, the transaction kind,
lifecycle state and status enums with their wire codecs and parse errors,
the transaction record, and `MigrationState` together with the methods that
merely read or reassemble it. The block was already contiguous, so this is a
pure move; the only edits are doc links that now cross a module boundary and
had to be spelled with a path.

`model` sits above the dependency-free planners it names in its types
(`denomination`, `preparation`, `scheduling`) and below everything else, so
the graph is acyclic. Every moved type is re-exported from `engine`, so the
paths downstream crates already use keep resolving unchanged.
